Send SIGINT to the correct process on stop
Due to changes in the way logging is implemented, the bash launched inside the scree session contains two subprocesses when logging is enabled. This confused the detection of the real user subprocess. As a consequence, this user subprocess never received a SIGINT and was kill in a not so nice fashion by terminating the screen session itself.

This commit adds a detection of the correct subprocess by grepping some vdemo-specific stuff in the process command line.