Correct screen sessions selection from title
The matching of screen session names to select them by VDEMO_title was
too greedy and resulted in false positives in case multiple configured
components hat the same prefix in their effective VDEMO_title (e.g.
through the use of the -t option). This commit matches all session names
including a whitespace character after the session name to ensure that
we have matched the whole name in the output of screen -ls.