Adds an autostop option -Q
When adding the -Q option to a component, the complete vdemo system will
now be automatically shut down in a clean fashion (stop all components,
gui exit) in case that respective component exits on its own. That means
if inotify is present and an autonomous termination of the screen
sessions is noticed, vdemo will exit. It will not exit, in cases the
user pressed the stop button in the gui.

An appropriate field to the gui was added so that the flag can be
changed at runtime.